refactor: isolate Discord media upload methods before routing repair
